build: release version 1.20.0 (#740)
At this point, the release of _Timefold Quickstarts_ is ready to be
published.

- Release branch has been created.
- Git tag has been published.

To finish the release of _Timefold Quickstarts_,
review and merge this PR to update the `stable` branch with new code.
Afterward, delete the branch that this PR is based on.
(Typically a button appears on this page once the PR is merged.)

Co-authored-by: Timefold Release Bot <release@timefold.ai>